+++ Project/src/EditItinScreen.js
@@ -8,6 +8,9 @@ import { Country, State, City }  from 'country-state-city';
 import { editItin, getMyItinsFromServer } from './actions/itin';
 import { TouchableOpacity } from 'react-native-gesture-handler';
 
+
+
+
 const EditItinScreen = ( {route, navigation} ) => {
 
   const dispatch = useDispatch();
@@ -41,6 +44,20 @@ const EditItinScreen = ( {route, navigation} ) => {
   const [stateList, setStateList] = useState( [] );
 
   const [cityList, setCityList] = useState( [] );
+
+  const handleSubmit = async () => {
+    setLoading(true);
+    await editItin(tid, jwt, uid, title, city, state, country, description, itinerary)(dispatch);
+    await getMyItinsFromServer(jwt, uid)(dispatch);
+    setLoading(false);
+    navigation.goBack();
+  }
+
+
+
+  
+
+  const myPlaces = useSelector(state => state.places);
 
 
   // const [stateDropDownDisabled, setStateDropDownDisabled] = useState(false);
@@ -87,14 +104,6 @@ const EditItinScreen = ( {route, navigation} ) => {
   }
 
 
-  const handleSubmit = async () => {
-    setLoading(true);
-    await editItin(tid, jwt, uid, title, city, state, country, description, itinerary)(dispatch);
-    await getMyItinsFromServer(jwt, uid)(dispatch);
-    setLoading(false);
-    navigation.goBack();
-  }
-
   const deleteDestination = (index) => {
     let newItin = [...itinerary];
     newItin.splice(index, 1); // 2nd parameter means remove one item only
@@ -103,17 +112,12 @@ const EditItinScreen = ( {route, navigation} ) => {
   }
 
   const addDestination = (destinationObj) => {
-    destinationObj = {
-      "extra_info": {},
-      "place_json": {
-        "name": "Destination",
-        "location": "New York",
-        "type": "School",
-        "url": "https://google.com/"
-      }
-    }
+    // const newObj = {
+    //   "extra_info": {},
+    //   "place_json": destinationObj
+    // }
     let newItin = [...itinerary, destinationObj];
-    setItinerary(newItin)
+    setItinerary(newItin);
   }
 
 
@@ -121,29 +125,33 @@ const EditItinScreen = ( {route, navigation} ) => {
     addDestination()
   }
 
-  const DestinationEditableComponent = ( {item, index} ) => {
+  const DestinationEditableComponent = ( {item, index, type} ) => {
     return (
-      <TouchableOpacity style={styles.placeContainer} onPress={ () => handlePlacePress(index) }>
+      <TouchableOpacity style={styles.placeContainer} onPress={ () => handlePlacePress(index, type, item) }>
         <Text style={{fontSize:25}}>{item.place_json.name}</Text>
         <Text>{item.place_json.address}</Text>
       </TouchableOpacity>
     )
   }
   
-  const handlePlacePress = (index) => {
-  
-    Alert.alert('Delete Destination', 'Delete Destination from your itinerary?', [
-      {
-        text: 'Delete',
-        onPress: () => deleteDestination(index),
-        style: 'destructive',
-      },
-      {
-        text: 'Cancel',
-        style: 'cancel',
-      },
-    ]);
-  
+  const handlePlacePress = (index, type, item) => {
+    if (type == "delete") {
+        Alert.alert('Delete Destination', 'Delete Destination from your itinerary?', [
+          {
+            text: 'Delete',
+            onPress: () => deleteDestination(index),
+            style: 'destructive',
+          },
+          {
+            text: 'Cancel',
+            style: 'cancel',
+          },
+        ]);
+    }
+    else {
+      addDestination(item)
+    }
+
   }
 
   return (
@@ -212,11 +220,18 @@ const EditItinScreen = ( {route, navigation} ) => {
 
       <Text style={styles.titleTextSecondary}>Destinations</Text>
       {
-        (itinerary.map) ?
-        (itinerary.map( (item, index) => <DestinationEditableComponent item={item} index={index} key={index}/> )):
+        (itinerary && itinerary.map) ?
+        (itinerary.map( (item, index) => <DestinationEditableComponent item={item} index={index} type={"delete"}/> )):
         <Text>Invalid.</Text>
       }
-        <Button title="+ Add Destination" onPress={handleAddDestinationPress}/>
+      <Text>● END OF ITINERARY</Text>
+      <Text style={{fontSize:21, fontWeight: "bold", marginTop: 5}}>Add a Saved Destination</Text>
+      {
+        (myPlaces && myPlaces.map) ?
+        (myPlaces.map( (item, index) => <DestinationEditableComponent item={item} index={index} type={"add"}/> )):
+        <Text>Invalid.</Text>
+      }
+        {/* <Button title="+ Add Destination" onPress={handleAddDestinationPress}/> */}
       {/* <Text>{JSON.stringify(itinerary)}</Text> */}
